Add analysis-collector plugin
Without this plugin the piperPipeline fails with:

java.lang.UnsupportedOperationException: no known implementation of interface
jenkins.tasks.SimpleBuildStep is named AnalysisPublisher.

The failure is triggered from checksPublishResults.groovy:108 which in turn
is triggered from piperPipelineStagePRVoting.groovy:102. This in turn gets called
from piperPipline stage "Pull-Request Voting"

By the way: there is a note on the README of the plugins source code repo
(https://github.com/jenkinsci/analysis-collector-plugin) stating that the plugin
is not supported anymore. But anyway we rely on the AnalysisPublisher class
provided by that plugin. We should resolve this, but as long as we use the
AnalysisPublisher plugin in the piperPipeline we should also have the plugin in
the cx server in order to avoid errors as mentioned above.