Build Jenkins master image on a schedule (#45)
This change replaces the automated build (and test) of Jenkins master on Docker hub. It also schedules a daily build and push (if smoke test succeeds) of the latest tag so we provide an up-to-date Jenkins master with updated plugins. 

The goals of this is to provide a more secure jenkins environment.

A secondary goal is to migrate off Docker hub as a build environment as it caused issues with non-triggered builds in the past.