Allow path to backup file in restore command
This change enables the user to provide the path (absolute or relative)
to the backup file, instead of just the backup filename.

Previously, ./cx-server restore only allowed the name of the backup
file as an argument. The backup file, however, does not reside in the
folder from where the cx-server script is usually executed. Instead,
the backup file resides in the folder backup/, making it difficult for
the user to determine the filename. It is easier for users to make use
of shell autocompletion, thus providing a path relative to the current
folder.