Fix release process (#71)
Currently the release process is broken because the GitHub api does not handle "get latest release" well if multiple tags point to the same commit, cf api docs https://developer.github.com/v3/repos/releases/#get-the-latest-release

Quote: 

> The latest release is the most recent non-prerelease, non-draft release, sorted by the created_at attribute. The created_at attribute is the date of the commit used for the release, and not the date when the release was drafted or published.

So if both `v10` and `v11` point to the same commit, `v10` is returned which breaks our process because it tries to release `v11` again. Still it makes sense to do those releases to rebuild the docker images and get the security updates in the base images.

The new way is to get the list of releases and manually "take the latest" which assumes the list is sorted (which it seems to be, but maybe that is also a false assumption). Additional sorting might be required, if this turns out to be an issue.